# Imporving speed by changing `wildcard` search process from `regular expression` match to `starts_with/ends_with` match