--- a/src/detections/rule/matchers.rs
+++ b/src/detections/rule/matchers.rs
@@ -2034,26 +2034,7 @@ mod tests {
     }
 
     #[test]
-    fn test_wildcard_converted_ends_with_notedetect() {
-        // ワイルドカード1文字を先頭に含む場合、ends_with相当のマッチ
-        let rule_str = r#"
-        enabled: true
-        detection:
-            selection:
-                Computer: '*-HOSTA'
-        details: 'command=%CommandLine%'
-        "#;
-
-        let record_json_str = r#"{
-            "Event": {"System": {"EventID": 4103, "Channel": "Security", "Computer": "A-HOST"}},
-            "Event_attributes": {"xmlns": "http://schemas.microsoft.com/win/2004/08/events/event"}
-        }"#;
-
-        check_select(rule_str, record_json_str, false);
-    }
-
-    #[test]
-    fn test_wildcard_converted_ends_with_shorter_val_notedetect() {
+    fn test_wildcard_converted_ends_with_shorter_val_notdetect() {
         // ワイルドカード1文字を先頭に含みかつ、比較対象文字のほうが文字数が少ない場合はマッチしない
         let rule_str = r#"
         enabled: true