@fukusuket 2023-01-23
@hitenkoku 
Thank you for your review🙇 I fixed the test code you commented :)