@fukusuket 2023-01-23
ベンチマークありがとうございます🙇引き続き正規表現マッチ減らせるところ調査いたします！